backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
alembic==1.13.2
psycopg[binary]==3.2.3
pydantic-settings==2.4.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
email-validator==2.2.0
